+++ src/App.tsx
@@ -1,5 +1,4 @@
 import Treege from "@/features/Treege";
-import "@/styles/globals.scss";
 import "@/config/i18n";
 
 const App = () => <Treege authToken="MyToken" endPoint="myEndPoint" />;