+++ src/components/Layouts/Main/Main.tsx
@@ -1,13 +1,20 @@
 import { Box } from "design-system-tracktor";
 import type { ReactNode } from "react";
-import styles from "./Main.module.scss";
+import colors from "@/styles/colors.module.scss";
 
 interface MainProps {
   children?: ReactNode;
 }
 
+const styles = {
+  box: {
+    border: `solid 1px ${colors.borderLight}`,
+    flexGrow: 1,
+  },
+};
+
 const Main = ({ children }: MainProps) => (
-  <Box className={styles.Box} component="main" role="tree">
+  <Box sx={styles.box} component="main" role="tree">
     {children}
   </Box>
 );