+++ src/components/UI/TreeModal/TreeModal.tsx
@@ -1,7 +1,7 @@
 import CloseRoundedIcon from "@mui/icons-material/CloseRounded";
 import { AppBar, Box, Dialog, IconButton, Slide, Toolbar, TransitionProps } from "design-system-tracktor";
 import { forwardRef, ReactElement, ReactNode, Ref } from "react";
-import styles from "./TreeModal.module.scss";
+import colors from "@/styles/colors.module.scss";
 
 interface TreeModalProps {
   children?: ReactNode;
@@ -9,6 +9,18 @@ interface TreeModalProps {
   open: boolean;
   onClose?(): void;
 }
+
+const styles = {
+  mainContainer: {
+    border: `solid 1px ${colors.borderLight}`,
+    height: " 100%",
+  },
+  toolbar: {
+    backgroundColor: colors.secondaryLight,
+    boxShadow: "none",
+    display: "flex",
+  },
+};
 
 const Transition = (
   { children, appear, in: inProps, onEnter, onExited, onFocus, role, tabIndex, timeout }: { children: ReactElement } & TransitionProps,
@@ -43,7 +55,7 @@ const TreeModal = ({ children, open, onClose, title }: TreeModalProps) => (
   >
     <>
       <AppBar position="sticky">
-        <Toolbar className={styles.Toolbar}>
+        <Toolbar sx={styles.toolbar}>
           <Box sx={{ flex: 1, ml: 2 }}>
             <h2>{title}</h2>
           </Box>
@@ -52,7 +64,7 @@ const TreeModal = ({ children, open, onClose, title }: TreeModalProps) => (
           </IconButton>
         </Toolbar>
       </AppBar>
-      <Box className={styles.MainContainer} m={2}>
+      <Box sx={styles.mainContainer} m={2}>
         {children}
       </Box>
     </>