+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # [Versions](https://github.com/Tracktor/treege/releases)
 
 ## v0.2.0
-
-- Add description
+- Documentation