+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v0.2.0
 - Documentation
 - Define global output vite config
+- Build script