+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # [Versions](https://github.com/Tracktor/treege/releases)
 
-## v0.1.0
-
-- Treege as library
+## v0.3.0
+- Fix style for library mode